+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Released: Upcoming
 - Added the ability to hand in default values to the create form
 - Added edit and delete buttons to the children table of the entities show page
 - CRUDMySQLData::listEntries handles null filter values now correctly
+- Fixed values are also returned in CRUDEntity::get() now
 
 ## 0.9.5
 Released: 2014-11-07